requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
click>=8.0.0
watchdog>=2.1.0 